+++ packages/css/src/components/form/form.stories.ts
@@ -6,6 +6,7 @@ import { formGroupCheckboxesTemplate } from './form-groups/checkboxes/form-group
 import { formGroupConfirmTemplate } from './form-groups/confirm/form-group-confirm.template';
 import { formGroupInputTemplate } from './form-groups/input/form-group-input.template';
 import { formGroupRadiosTemplate } from './form-groups/radios/form-group-radios.template';
+import { formGroupTextareaTemplate } from './form-groups/textarea/form-group-textarea.template';
 
 import { formTemplate } from './form.template';
 import readme from './readme.md';
@@ -22,6 +23,7 @@ storiesOfForm(
     formGroupCheckboxesTemplate,
     formGroupConfirmTemplate,
     formGroupInputTemplate,
-    formGroupRadiosTemplate
+    formGroupRadiosTemplate,
+    formGroupTextareaTemplate
   }
 );