+++ packages/sources/src/components/form/form.stories.ts
@@ -8,11 +8,14 @@ import { checkboxesContent } from './content/checkboxes.content';
 import { confirmContent } from './content/confirm.content';
 import { inputContent } from './content/input.content';
 import { radiosContent } from './content/radios.content';
-import { Form, FormGroupCheckboxes, FormGroupConfirm, FormGroupInput, FormGroupRadios, FormGroupInputDate } from './form.models';
+import { textareaContent } from './content/textarea.content';
+import { Form, FormGroupCheckboxes, FormGroupConfirm, FormGroupInput, FormGroupRadios, FormGroupInputDate, FormGroupTextarea } from './form.models';
 import { formGroupCheckboxesArgsMapper, formGroupCheckboxesArgTypes } from './form-groups/checkboxes/form-group-checkboxes.args';
 import { formGroupConfirmArgsMapper, formGroupConfirmArgTypes } from './form-groups/confirm/form-group-confirm.args';
 import { formGroupInputArgsMapper, formGroupInputArgTypes } from './form-groups/input/form-group-input.args';
 import { formGroupRadiosArgsMapper, formGroupRadiosArgTypes } from './form-groups/radios/form-group-radios.args';
+import { formGroupTextareaArgsMapper, formGroupTextareaArgTypes } from './form-groups/textarea/form-group-textarea.args';
+
 
 export interface FormParameters<TemplateFnReturnType> {
   formTemplate: (form: Form<TemplateFnReturnType>) => TemplateFnReturnType;
@@ -21,6 +24,7 @@ export interface FormParameters<TemplateFnReturnType> {
   formGroupConfirmTemplate: (formGroupConfirm: FormGroupConfirm<TemplateFnReturnType>) => TemplateFnReturnType;
   formGroupInputTemplate: (formGroupInput: FormGroupInput<TemplateFnReturnType> | FormGroupInputDate<TemplateFnReturnType>) => TemplateFnReturnType;
   formGroupRadiosTemplate: (formGroupRadios: FormGroupRadios<TemplateFnReturnType>) => TemplateFnReturnType;
+  formGroupTextareaTemplate: (formGroupTextarea: FormGroupTextarea<TemplateFnReturnType>) => TemplateFnReturnType;
 }
 
 export function storiesOfForm<TemplateFnReturnType>(
@@ -35,7 +39,8 @@ export function storiesOfForm<TemplateFnReturnType>(
     formGroupCheckboxesTemplate,
     formGroupConfirmTemplate,
     formGroupInputTemplate,
-    formGroupRadiosTemplate
+    formGroupRadiosTemplate,
+    formGroupTextareaTemplate
   }: FormParameters<TemplateFnReturnType>
 ) {
   storiesOf('Form/form', mainModule)
@@ -95,4 +100,15 @@ export function storiesOfForm<TemplateFnReturnType>(
     .addDecorator(formGroupDecorator)
     .add('radios', bindTemplate(formGroupRadiosArgsMapper, formGroupRadiosTemplate)
   );
-}
+
+  storiesOf('Form/groups/textarea', mainModule)
+    .addParameters({
+      argTypes: formGroupTextareaArgTypes,
+      args: textareaContent,
+      docs: {
+        page: readme
+      }
+    })
+    .addDecorator(formGroupDecorator)
+    .add('textarea', bindTemplate(formGroupTextareaArgsMapper, formGroupTextareaTemplate)
+  );}