+++ packages/sources/src/components/form/models/form-group-textarea.model.ts
@@ -1,0 +1,11 @@
+import { Icon } from '../../..';
+
+import { FormGroupBase } from './form-group.base-model';
+
+export interface FormGroupTextarea<TemplateFnReturnType> extends FormGroupBase<TemplateFnReturnType> {
+  group: 'textarea',
+  value?: string;
+  placeholder?: string;
+  rows?: number;
+  feedback?: Icon;
+}